PSLR submission pages quick fixes
